requirements.txt
numpy==1.21.6
opencv-contrib-python==4.7.0.72
pandas==1.1.5
Pillow==9.5.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
